Update NGINX Nextcloud configuration.
The recommended NGINX configuration in the latest Nextcloud
documentation was once again updated.  Update the NGINX Nextcloud
configuration overlay, accordingly.

The update does the following:
  * simplifies rules for well-known locations and asset handling
  * adds a set of default cache control policies (expiry)
  * adds support for Microsoft DAV clients

The updated configuration is a verbatim copy of the recommended
configuration with modifications specific to the FreeNAS/TrueNAS
Nextcloud plugin, i.e. no SSL/TLS configuration, a different root
path, a larger maximum file upload size, and enabling of strict
transport security.  This is consistent with earlier versions of the
Nextcloud plugin NGINX configuration.